+++ admins/manage-users.php
@@ -1,6 +1,7 @@
 <?php
 include('./partials/header.php');
-//fetching users from db but not current
+
+// Fetching users from db but not current
 $current_admin_id = $_SESSION['user-id'];
 $query = "SELECT * FROM users WHERE NOT id=$current_admin_id";
 $users = mysqli_query($connection, $query);
@@ -9,40 +10,31 @@ $users = mysqli_query($connection, $query);
 <section class="dashboard">
     <?php if (isset($_SESSION['add-user-success'])) : ?>
         <div class="alert__message success container">
-
             <p><?= $_SESSION['add-user-success'];
-                unset($_SESSION['add-user-success']);
-                ?></p>
+                unset($_SESSION['add-user-success']); ?></p>
         </div>
     <?php elseif (isset($_SESSION['edit-user-success'])) : ?>
         <div class="alert__message success container">
-
             <p><?= $_SESSION['edit-user-success'];
-                unset($_SESSION['edit-user-success']);
-                ?></p>
+                unset($_SESSION['edit-user-success']); ?></p>
         </div>
     <?php elseif (isset($_SESSION['edit-user'])) : ?>
         <div class="alert__message error container">
-
             <p><?= $_SESSION['edit-user'];
-                unset($_SESSION['edit-user']);
-                ?></p>
+                unset($_SESSION['edit-user']); ?></p>
         </div>
-        <?php elseif (isset($_SESSION['delete-user-success'])) : ?>
+    <?php elseif (isset($_SESSION['delete-user-success'])) : ?>
         <div class="alert__message success container">
-
             <p><?= $_SESSION['delete-user-success'];
-                unset($_SESSION['delete-user-success']);
-                ?></p>
+                unset($_SESSION['delete-user-success']); ?></p>
         </div>
     <?php elseif (isset($_SESSION['delete-user'])) : ?>
         <div class="alert__message error container">
-
             <p><?= $_SESSION['delete-user'];
-                unset($_SESSION['delete-user']);
-                ?></p>
+                unset($_SESSION['delete-user']); ?></p>
         </div>
     <?php endif; ?>
+
     <div class="container dashboard__container">
         <button id="show__sidebar-btn" class="sidebar__toggle"><i class="uil uil-angle-left-b"></i></button>
         <button id="hide__sidebar-btn" class="sidebar__toggle"><i class="uil uil-angle-right-b"></i></button>
@@ -62,7 +54,6 @@ $users = mysqli_query($connection, $query);
                     </a>
                 </li>
                 <?php if (isset($_SESSION['user_is_admin'])) : ?>
-
                     <li>
                         <a href="add-user.php">
                             <i class="uil uil-user-plus"></i>
@@ -90,37 +81,41 @@ $users = mysqli_query($connection, $query);
                 <?php endif ?>
             </ul>
         </aside>
+
         <main>
             <h2>Manage Users</h2>
-            <table>
-                <thead>
-                    <tr>
-                        <th>Name</th>
-                        <th>Username</th>
-                        <th>Edit</th>
-                        <th>Delete</th>
-                        <th>Admin</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    <?php while ($user = mysqli_fetch_assoc($users)) : ?>
+            <?php if (mysqli_num_rows($users) > 0) : ?>
+                <table>
+                    <thead>
                         <tr>
-                            <td><?= "{$user['firstname']} {$user['lastname']}" ?></td>
-                            <td><?= $user['email'] ?></td>
-                            <td> <a href="<?= ROOT_URL ?>admins/edit-user.php?id=<?= $user['id'] ?>" class="btn sm">Edit</a></td>
-                            <td> <a href="<?= ROOT_URL ?>admins/delete-user.php?id=<?= $user['id'] ?>" class="btn sm danger">Delete</a></td>
-                            <td><?= $user['is_admin'] ? 'Yes' : 'No' ?></td>
+                            <th>Name</th>
+                            <th>Username</th>
+                            <th>Edit</th>
+                            <th>Delete</th>
+                            <th>Admin</th>
                         </tr>
-                    <?php endwhile ?>
-                </tbody>
-            </table>
+                    </thead>
+                    <tbody>
+                        <?php while ($user = mysqli_fetch_assoc($users)) : ?>
+                            <tr>
+                                <td><?= "{$user['firstname']} {$user['lastname']}" ?></td>
+                                <td><?= $user['email'] ?></td>
+                                <td><a href="<?= ROOT_URL ?>admins/edit-user.php?id=<?= $user['id'] ?>" class="btn sm">Edit</a></td>
+                                <td><a href="<?= ROOT_URL ?>admins/delete-user.php?id=<?= $user['id'] ?>" class="btn sm danger">Delete</a></td>
+                                <td><?= $user['is_admin'] ? 'Yes' : 'No' ?></td>
+                            </tr>
+                        <?php endwhile ?>
+                    </tbody>
+                </table>
+            <?php else : ?>
+                <div class="alert__message error"><?= "No users found." ?></div>
+            <?php endif; ?>
         </main>
-
 
     </div>
 </section>
 
-<script src="..\js\main.js"></script>
+<script src="../js/main.js"></script>
 </body>
 
 </html>
